MSI-2293: Create Orders With Simple Product In Multi Stock Mode automation.
